Worked on feedbacks
