+++ package.json
@@ -7,12 +7,14 @@
 		"node": "8.11.2"
 	},
 	"scripts": {
-		"start": "nodemon server.js --exec babel-node --presets es2015",
-		"parcelTable": "node models/parcelsdb createTables",
-		"userTable": "node models/userdb createTables",
+		"start": "nodemon ./server/server.js --exec babel-node --presets es2015",
+		"parcelTable": "node ./server/models/parcelsdb createTables",
+		"userTable": "node ./server/models/userdb createTables",
 		"startdb": "psql -U postgres -h127.0.0.1 sendit_db",
-		"test": "nyc mocha --exit --timeout 30000 --compilers js:babel-core/register ./test/*.js",
-		"coverage": "nyc report --reporter=text-lcov | coveralls"
+		"test": "nyc mocha --exit --timeout 30000 --compilers js:babel-core/register ./server/test/*.js",
+		"coverage": "nyc report --reporter=text-lcov | coveralls",
+		"lint": "./node_modules/.bin/eslint ./server --config .eslintrc.json",
+		"lint-fix": "./node_modules/.bin/eslint ./server --fix .eslintrc.json"
 	},
 	"repository": {
 		"type": "git",
@@ -32,12 +34,9 @@
 		"babel-preset-env": "^1.7.0",
 		"babel-preset-es2015": "6.24.1",
 		"babel-register": "6.26.0",
-		"body-parser": "1.18.3",
-		"chai": "^4.2.0",
-		"chai-http": "4.2.0",
 		"coveralls": "3.0.2",
 		"cross-env": "^5.2.0",
-		"eslint": "^5.8.0",
+		"eslint": "^5.9.0",
 		"eslint-config-airbnb": "17.1.0",
 		"eslint-config-airbnb-base": "13.1.0",
 		"eslint-plugin-import": "2.14.0",
@@ -53,6 +52,9 @@
 	"dependencies": {
 		"bcrypt": "^3.0.2",
 		"bluebird": "^3.5.3",
+		"body-parser": "^1.18.3",
+		"chai": "^4.2.0",
+		"chai-http": "^4.2.0",
 		"dotenv": "^6.1.0",
 		"express": "4.16.4",
 		"express-validator": "5.3.0",
