+++ server/server.js
@@ -5,22 +5,22 @@ import router from './routes/route';
 const app = express();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({
-	extended: true,
+  extended: false,
 }));
 
 
 router.get('/', (req, res) => res.status(200).json({
-	success: true,
-	message: 'Send It API',
+  success: true,
+  message: 'Send It API',
 }));
 
 
-app.use(router);
+app.use('/api/v1', router);
 
 const PORT = 3000;
 
 app.listen(PORT, () => {
-	console.log('Listening...');
+  console.log('Listening...');
 });
 
 export default app;
